+++ packages/dev-iroha-bins/src/types.ts
@@ -10,12 +10,3 @@ export enum KnownBinaries {
 }
 
 export type BinaryNameMap = { [K in KnownBinaries]: string }
-
-export interface Config {
-  git: {
-    repo: string
-    branch?: string
-    revision?: string
-  }
-  binaries: BinaryNameMap
-}